Context._options defaults to argparse.Namespace
Updated default value of `Context._options` from `None` to `Namespace`.
This makes member access operate the same regardless of where used and
to remove the need for None checking. It also provides a hint to plugin
developers for how to deserialize options from a dict if they had to.
Opted for `argparse.Namespace` instead of `types.SimpleNamespace` to
support Python 2.7.