Make Dispatcher have all plugins by default
This used to be the behavior, and then
b5499c7dc5b300462f3ce1c2a3d9b7a76233b39b changed this, breaking some
plugins (some of which subsequently implemented workarounds). This patch
restores the behavior, so that if a Dispatcher is constructed without
explicitly passing in plugins, it has access to all plugins.

This change seems safe, in that it's unlikely that any plugins were
relying on the behavior between b5499c7dc5b300462f3ce1c2a3d9b7a76233b39b
and this patch, where a Dispatcher without an explicit plugin list
behaved as if there were no plugins---because such a Dispatcher is not
very useful!