Add Python 3.14 to CI
This patch updates the testing code to not wrap `copy_file_range` with a
check of the destination file: because this function works on file
descriptors rather than paths, it's not possible to check the path of
the destination. On Linux / Python 3.14, this function is called by
`shutil.copyfile`, which we do wrap, so the checking happens at that
level.